fix(cookies): adding apis in Response
